Split out model default utils
